requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg[binary]>=3.2.2
python-multipart==0.0.6
httpx==0.25.1
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
